+++ packages/core/src/render3/instructions/all.ts
@@ -25,6 +25,7 @@
  *
  * Jira Issue = FW-1184
  */
+export * from '../../defer/instructions';
 export * from './advance';
 export * from './attribute';
 export * from './attribute_interpolation';
@@ -32,7 +33,6 @@ export * from './change_detection';
 export * from './class_map_interpolation';
 export * from './component_instance';
 export * from './control_flow';
-export * from '../../defer/instructions';
 export * from './di';
 export * from './di_attr';
 export * from './element';
@@ -47,6 +47,7 @@ export * from './next_context';
 export * from './projection';
 export * from './property';
 export * from './property_interpolation';
+export * from './queries';
 export * from './storage';
 export * from './style_map_interpolation';
 export * from './style_prop_interpolation';
